+++ src/Icarus.Domain/Entities/Category.cs
@@ -1,7 +1,7 @@
 ﻿using Icarus.Domain.Commons;
 
 namespace Icarus.Domain.Entities;
-public class Category : Auditable<short>
+public class Category : Auditable<long>
 {
     public string Name { get; set; }
     public ICollection<DepartmentCategory> DepartmentCategories { get; set; }
